% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchAvg}
\alias{catchAvg}
\title{averaged catch in the North Sea 2014 - 2019}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
catchAvg
}
\description{
....
}
\keyword{datasets}
