% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fastOptim}
\alias{fastOptim}
\title{Wrapper for optimParallel}
\usage{
fastOptim(
  params,
  vary,
  vary_df,
  errorFun,
  data_type = "biomass_observed",
  effort = 0,
  time_series = NULL,
  spareCores = 1
)
}
\arguments{
\item{params}{A mizer params object}

\item{vary}{Vector containing the values of the parameters to vary}

\item{vary_df}{Dataframe used to fill the params object with \code{vary}. Its dimension
should be number of different parameters x 6. Its column are \code{name}, \code{length}
for the number of values associated to the parameter, \code{lower} and \code{upper} for
the lower and upper boundaries to explore with \code{optimparallel}, \code{slot} for the
slot name in the params object where the parameter is stored and \code{unit} if the
parameter is not stored in \code{vary} in linear unit.}

\item{errorFun}{What error function should be used}

\item{data_type}{Data getError is compared to. Can be yield_observed or
biomass_observed. A column named the same way must be present in the params
object, which contain the empirical data to compare to.}

\item{effort}{Effort values to pass to the \code{project} function when using
\code{getErrorCustom}. Default is 0 (no fishing effort).}

\item{time_series}{If the result of the projected biomass or yield must be
compared to a time series of value instead of a unique value found in the
"observed" column of species_params, the user can supply a matrix of values
through this argument. The dimension should be the same as \code{getBiomass} output
(time x species). Default is NULL}

\item{spareCores}{A numeric value to tell optimparallel how many cores should
be left unused. Default is one.}
}
\description{
The function sets up the parallel environment to run optimParallel
on a params object
}
