% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fmat}
\alias{Fmat}
\title{ICES fisheries mortality}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 55 columns.
}
\usage{
Fmat
}
\description{
....
}
\keyword{datasets}
