% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FmatWeightedInterpolated}
\alias{FmatWeightedInterpolated}
\title{ICES fisheries mortality processed
On the ICES database, some species have multiple entries and some have none.
Fisheries mortality is weighted by the species abundance to be able to take into account multiple database entries
Gurnard data is extrapolated from whiting (same gear and physiology)
....}
\format{
An object of class \code{data.frame} with 74 rows and 13 columns.
}
\usage{
FmatWeightedInterpolated
}
\description{
ICES fisheries mortality processed
On the ICES database, some species have multiple entries and some have none.
Fisheries mortality is weighted by the species abundance to be able to take into account multiple database entries
Gurnard data is extrapolated from whiting (same gear and physiology)
....
}
\keyword{datasets}
