% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{g_legend}
\alias{g_legend}
\title{file that should be sourced at the beginning of every .Rmd
Contains libraries and functions}
\usage{
g_legend(a.gplot)
}
\description{
file that should be sourced at the beginning of every .Rmd
Contains libraries and functions
}
