% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getBiomassFrame2}
\alias{getBiomassFrame2}
\title{Plot ....}
\usage{
getBiomassFrame2(
  sim,
  species = dimnames(sim@n)$sp[!is.na(sim@params@A)],
  min_w = NULL,
  start_time = as.numeric(dimnames(sim@n)[[1]][1]),
  end_time = as.numeric(dimnames(sim@n)[[1]][dim(sim@n)[1]]),
  ylim = c(NA, NA),
  total = FALSE,
  ...
)
}
\description{
Plot ....
}
