% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getDietComp}
\alias{getDietComp}
\title{Get the diet composition}
\usage{
getDietComp(sim)
}
\arguments{
\item{sim}{An object of class \linkS4class{MizerSim}}
}
\value{
An array (predator species x predator size x
(prey species + resource) x prey size)
}
\description{
The diet \eqn{D_{ij}(w, w_p)} is the prey biomass density rate for a predator of
species \eqn{i} and weight \eqn{w}, resolved by prey species \eqn{j} and prey
size \eqn{w_p}. It is calculated from the predation kernel \eqn{\phi(w, w_p)},
the search volume \eqn{\gamma_i(w)}, the feeding level \eqn{f_i(w)}, the
species interaction matrix \eqn{\theta_{ij}} and the prey abundance density
\eqn{N_j(w)}:
\deqn{
D_{ij}(w, w_p) = (1-f_i(w)) \gamma_i(w) \theta_{ij} N_j(w_p)
\phi_i(w, w_p) w_p.
}
The prey index \eqn{j} can run over all species and the resource. The returned
values have units of 1/year.
}
\details{
The total rate \eqn{D_{ij}(w)} at which a predator of species \eqn{i}
and size \eqn{w} consumes biomass from prey species \eqn{j} is
obtained by integrating over prey sizes:
\deqn{
D_{ij}(w) = \int D_{ij}(w, w_p) dw_p.
}
This aggregated diet can also be obtained directly from the \code{getDiet()} function.
}
