% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getErrorCustom}
\alias{getErrorCustom}
\title{Improving getError}
\usage{
getErrorCustom(
  vary,
  vary_df,
  params,
  data_type = "biomass_observed",
  tol = 0.001,
  timetorun = 10
)
}
\arguments{
\item{vary}{Vector containing the values of the parameters to vary}

\item{vary_df}{Dataframe used to fill the params object with \code{vary}. Its dimension
should be number of different parameters x 6. Its column are \code{name}, \code{length}
for the number of values associated to the parameter, \code{lower} and \code{upper} for
the lower and upper boundaries to explore with \code{optimparallel}, \code{slot} for the
slot name in the params object where the parameter is stored and \code{unit} if the
parameter is not stored in \code{vary} in linear unit.}

\item{params}{A mizer params object}

\item{data_type}{Data getError is compared to. Can be yield_observed or
biomass_observed. A column named the same way must be present in the params
object, which contain the empirical data to compare to.}

\item{tol}{The simulation stops when the relative change in the egg production
RDI over t_per years is less than tol for every species. See \code{projectToSteadty}
documentation}

\item{timetorun}{\code{t_max} for the \code{project} function. Lower value will yield
faster run but to the cost of accuracy. Default is 10.}
}
\description{
This function is a more versatile version of \code{getError} that takes
any parameter from the species_params slot (currently but may be extended). To
work in tandem with optimparallel, the information about the parameters to be
varied is divided into two arguments: \code{vary} which must be numeric and \code{vary_df}
which is used to put the different values from vary in the right spots.
}
