% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTM0_sim}
\alias{HTM0_sim}
\title{HTM calibrated simulation from HTM3 (ask Julia for more details)
Used in HTM0}
\format{
An object of class \code{MizerSim} of length 1.
}
\usage{
HTM0_sim
}
\description{
....
}
\keyword{datasets}
