% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTM1_sim}
\alias{HTM1_sim}
\title{HTM1 final simulation saved (sim_guessed)
Used at the beginning of HTM2}
\format{
An object of class \code{MizerSim} of length 1.
}
\usage{
HTM1_sim
}
\description{
....
}
\keyword{datasets}
