% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTM2_Fmsy1}
\alias{HTM2_Fmsy1}
\title{HTM2 Fmsy plot file (step 2 - Fmsy1 plot_dat)
plotFmsy() runs slowly so saving the result for fast kniting}
\format{
An object of class \code{data.frame} with 590 rows and 3 columns.
}
\usage{
HTM2_Fmsy1
}
\description{
....
}
\keyword{datasets}
