% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTM2_optimParallel_Rmax2}
\alias{HTM2_optimParallel_Rmax2}
\title{HTM2 optim parallel output (step 2 - optim_loop)
The Rmax vector output of optim parallel is saved for fast knitting
....}
\format{
An object of class \code{list} of length 6.
}
\usage{
HTM2_optimParallel_Rmax2
}
\description{
HTM2 optim parallel output (step 2 - optim_loop)
The Rmax vector output of optim parallel is saved for fast knitting
....
}
\keyword{datasets}
