% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTM2_sim_optim2}
\alias{HTM2_sim_optim2}
\title{HTM2 saved simulation (step 2 - tweaked erepro)
Used as checkpoint/kick start if .rmd are not run from the beginning
....}
\format{
An object of class \code{MizerSim} of length 1.
}
\usage{
HTM2_sim_optim2
}
\description{
HTM2 saved simulation (step 2 - tweaked erepro)
Used as checkpoint/kick start if .rmd are not run from the beginning
....
}
\keyword{datasets}
