% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsParams}
\alias{nsParams}
\title{Species specific parameters for the North Sea}
\format{
An object of class \code{data.frame} with 12 rows and 16 columns.
}
\usage{
nsParams
}
\description{
....
}
\keyword{datasets}
