% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotFeedingLevel2}
\alias{plotFeedingLevel2}
\title{Plot ....}
\usage{
plotFeedingLevel2(
  object,
  species = NULL,
  time_range,
  highlight = NULL,
  all.sizes = FALSE,
  include_critical = FALSE,
  return_data = FALSE,
  ...
)
}
\description{
Plot ....
}
