% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPPMR}
\alias{plotPPMR}
\title{Plot showing emergent PPMR}
\usage{
plotPPMR(sim, return_data = FALSE)
}
\arguments{
\item{sim}{An object of class MizerSim}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}
}
\description{
Yet another plot
}
