% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRdiVsRdd}
\alias{plotRdiVsRdd}
\title{Plot density independent versus density dependent reproduction rate}
\usage{
plotRdiVsRdd(object, return_data = FALSE)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}
}
\description{
yet another plot
}
