% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot ....}
\usage{
plotSummary(
  x,
  y,
  power = 1,
  wlim = c(0.001, NA),
  short = F,
  save_it = FALSE,
  name_save = NULL,
  ...
)
}
\description{
Plot ....
}
