% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{shiny_fmsy}
\alias{shiny_fmsy}
\title{shiny app to tweak erepro per species and see the effects on Fmsy}
\usage{
shiny_fmsy(params, dat)
}
\description{
shiny app to tweak erepro per species and see the effects on Fmsy
}
