% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{shiny_gamma}
\alias{shiny_gamma}
\title{shiny app to tweak gamma per species and see the effects on growth}
\usage{
shiny_gamma(params)
}
\description{
shiny app to tweak gamma per species and see the effects on growth
}
