% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{shiny_kappa}
\alias{shiny_kappa}
\title{shiny app to tweak gamma per species and see the effects on growth}
\usage{
shiny_kappa(params, dat = NULL)
}
\description{
shiny app to tweak gamma per species and see the effects on growth
}
