% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tutorial}
\alias{tutorial}
\title{file that should be sourced at the beginning of every .Rmd
Contains libraries and functions}
\usage{
tutorial()
}
\description{
Open tutorials
}
\details{
Opens the rmarkdown file of a tutorial in RStudio so that the user can
play with the code.
}
