% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{g_legend}
\alias{g_legend}
\title{Create ggplot legend object}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot object}
}
\description{
Function that extracts the legend from ggplot object to use it in grid plots
}
