% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getError}
\alias{getError}
\title{compare outputs}
\usage{
getError(vary, params, dat, data_type = "catch", tol = 0.1, timetorun = 10)
}
\description{
compare outputs
}
\examples{
using n . w . dw so g per year per volume (i.e. North Sea since kappa is set up this way).
The data are in tonnes per year so converting to tonnes.
}
