% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interNS}
\alias{interNS}
\title{Interaction matrix for the North Sea}
\format{
An object of class \code{matrix} with 12 rows and 12 columns.
}
\usage{
interNS
}
\description{
....
}
\keyword{datasets}
