% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerHowTo-package.R
\docType{package}
\name{mizerHowTo-package}
\alias{mizerHowTo}
\alias{mizerHowTo-package}
\title{mizerHowTo}
\description{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gustavdelius/mizerHowTo}
  \item Report bugs at \url{https://github.com/gustavdelius/mizerHowTo/issues}
}

}
