% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotCalibration}
\alias{plotCalibration}
\title{Summary function bundling helpful plots during calibration.}
\usage{
plotCalibration(
  sim,
  catch_dat = NULL,
  stage = 1,
  wlim = c(0.1, NA),
  power = 1,
  effortRes = 10
)
}
\arguments{
\item{sim}{An object of class MizerSim}

\item{stage}{Numeric values which determines the function's output. Range from 1 to 3.
Default is 1.}

\item{power}{The abundance is plotted as the number density times the weight raised to power.
The default power = 1 gives the biomass density, whereas power = 2 gives the biomass density
with respect to logarithmic size bins.}
}
\description{
Stage 1 shows size spectra, biomass and predicted versus observed yield
Stage 2 shows the pannel of fisheries effort versus yield
Stage 3 shows the pannel of growth curves
}
