% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotGrowthCurves2}
\alias{plotGrowthCurves2}
\title{Plot ....}
\usage{
plotGrowthCurves2(
  object,
  species = NULL,
  max_age = 20,
  percentage = FALSE,
  species_panel = FALSE,
  highlight = NULL,
  returnData = F
)
}
\description{
Plot ....
}
