% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPredObsYield}
\alias{plotPredObsYield}
\title{A function that plots the predicted yield versus observed yield.}
\usage{
plotPredObsYield(sim, dat, returnData = FALSE)
}
\arguments{
\item{sim}{An object of class MizerSim}

\item{dat}{A dataframe containing the observed yield values}
}
\description{
A function that plots the predicted yield versus observed yield.
}
