% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotSummary}
\alias{plotSummary}
\title{Summary plot displaying 8 differents plots useful to assess the state of
a mizerSim object.}
\usage{
plotSummary(
  sim,
  power = 1,
  save_it = FALSE,
  name_save = NULL,
  font_size = 9,
  ...
)
}
\arguments{
\item{sim}{An object of class MizerSim}

\item{power}{The abundance is plotted as the number density times the weight raised to power.
The default power = 1 gives the biomass density, whereas power = 2 gives the biomass density
with respect to logarithmic size bins.}

\item{save_it}{Boleean value that determines whether to save the output of the function or not.
Default is FALSE}

\item{name_save}{Character string to give a specific name if saving the plot. Default is NULL}

\item{font_size}{Determines the font size of the labels and ticks. Default is 9.}
}
\description{
Left column (top to bottom) is: size spectrum, feeding level, predation
and fishing mortality on linear scale, predation and fishing mortality
on a log scale
Right column (top to bottom) is: predicted catches and biomass, RDD/RDI,
biomass, PPMR per size of predator (work in progress)
}
